services/auth/requirements.txt
# Core Django
Django>=5.1.0,<5.2.0
psycopg2-binary>=2.9.0

# Django REST Framework & Authentication
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
djoser>=2.2.0

# Image Processing
Pillow>=10.0.0

# Environment Variables
python-decouple>=3.8

# API Documentation
drf-yasg>=1.21.0

# CORS Headers
django-cors-headers>=4.3.0

# Filtering
django-filter>=23.0

# Server
gunicorn>=21.2.0
